Portfolio/requirements.txt
Pillow==9.2.0
requests==2.28.1
streamlit==1.12.2
streamlit_lottie==0.0.3
